feat(mp24): S-MP18t — patch ConvoView max() + restore ConvoBox/Screen
ConvoView.cpp had two max(int_literal, int32_t_expr) call sites
that fail under modern C++ STL because std::max<T>(T, T) can't
deduce T from mismatched argument types (and arduino-esp32 3.x
explicitly #undefs the Arduino max macro before STL include).

Lines 22 + 40 — patched both args to (int32_t):

  -    this->startIndex = max(0, (int32_t) startIndex);
  +    this->startIndex = max((int32_t) 0, (int32_t) startIndex);

  -    int32_t index = max(0, (int32_t) convo.messages.size() - Count);
  +    int32_t index = max((int32_t) 0, (int32_t) convo.messages.size() - Count);

Semantics unchanged — both args now agree on type so the macro
or std::max resolve identically. Two-character source delta.

Restored to chatter_app SRCS:
  Interface/ConvoView.cpp     (the patched file)
  Elements/ConvoBox.cpp       (constructs ConvoView; needs above)
  Screens/ConvoScreen.cpp     (uses ConvoBox)

Compile-coverage milestone reached: chatter_app now covers
every screen in the messaging-flow path, plus the entire
games / settings / utility / call ecosystem. ~22k LOC of
upstream code in the build.

+++ src/Interface/ConvoView.cpp
@@ -19,7 +19,7 @@ void ConvoView::load(size_t startIndex){
 	if(startIndex + Count >= convo.messages.size()){
 		startIndex = convo.messages.size() - Count;
 	}
-	this->startIndex = max(0, (int32_t) startIndex);
+	this->startIndex = max((int32_t) 0, (int32_t) startIndex);
 
 	latest = this->startIndex + Count >= convo.messages.size();
 
@@ -37,7 +37,7 @@ void ConvoView::loadLatest(){
 	Convo convo = Storage.Convos.get(convoUID);
 	if(convo.uid == 0) return;
 
-	int32_t index = max(0, (int32_t) convo.messages.size() - Count);
+	int32_t index = max((int32_t) 0, (int32_t) convo.messages.size() - Count);
 	load(index);
 }
